+++ CHANGELOG.md
@@ -29,8 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   a sender to send XRP to the specified XRPL account that has enabled Deposit Authorization.
   (See https://xrpl.org/depositpreauth.html)
 
-- A new method `authorizeSendingAccount` is added to `XrpClient` which authorizes
-  senders to send XRP to the specified XRPL account.
+- A new method `authorizeSendingAccount` is added to `XrpClient` which authorizes a
+  sender to send XRP to the specified XRPL account that has enabled Deposit Authorization.
   (See https://xrpl.org/depositpreauth.html)
 
 ## 5.1.2 - 2020-10-02